requirements.txt
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.3
python-dotenv==1.0.0
celery==5.2.7
redis==4.5.4
peewee==3.16.2
psycopg2==2.9.6
python-logstash-async==2.5.0
flask-admin==1.6.1
wtf-peewee==3.0.4
gunicorn==20.1.0

